Ships Beta 6

+++ src/main/java/org/ships/movement/autopilot/scheduler/FallExecutor.java
@@ -7,8 +7,6 @@ import org.core.world.boss.ServerBossBar;
 import org.ships.config.configuration.ShipsConfig;
 import org.ships.exceptions.MoveException;
 import org.ships.movement.MovementContext;
-import org.ships.movement.result.AbstractFailedMovement;
-import org.ships.movement.result.FailedMovement;
 import org.ships.movement.result.MovementResult;
 import org.ships.plugin.ShipsPlugin;
 import org.ships.vessel.common.assits.Fallable;
@@ -18,7 +16,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.util.ConcurrentModificationException;
-import java.util.concurrent.TimeUnit;
 
 public class FallExecutor implements Runnable {
 
@@ -34,34 +31,35 @@ public class FallExecutor implements Runnable {
                         v.getEntities().stream().filter(e -> e instanceof LivePlayer).forEach(e -> bar.register((LivePlayer) e));
                         context.setBar(bar);
                     }
-                    try {
-                        v.moveTowards(0, -(ShipsPlugin.getPlugin().getConfig().getFallingSpeed()), 0, context);
-                    } catch (MoveException e) {
-                        context.getBar().ifPresent(ServerBossBar::deregisterPlayers);
-                        if (e.getMovement().getResult().equals(MovementResult.COLLIDE_DETECTED)) {
-                            if (v instanceof FileBasedVessel) {
-                                File file = ((FileBasedVessel) v).getFile();
-                                try {
-                                    Files.delete(file.toPath());
-                                } catch (IOException e2) {
-                                    e.printStackTrace();
+                    v.moveTowards(0, -(ShipsPlugin.getPlugin().getConfig().getFallingSpeed()), 0, context, exc -> {
+                        if (exc instanceof MoveException) {
+                            MoveException e = (MoveException) exc;
+                            context.getBar().ifPresent(ServerBossBar::deregisterPlayers);
+                            if (e.getMovement().getResult().equals(MovementResult.COLLIDE_DETECTED)) {
+                                if (v instanceof FileBasedVessel) {
+                                    File file = ((FileBasedVessel) v).getFile();
+                                    try {
+                                        Files.delete(file.toPath());
+                                    } catch (IOException e2) {
+                                        e.printStackTrace();
+                                    }
                                 }
+                                ShipsPlugin.getPlugin().unregisterVessel(v);
+                                v.getPosition().destroy();
                             }
-                            ShipsPlugin.getPlugin().unregisterVessel(v);
-                            v.getPosition().destroy();
+                        } else {
+                            context.getBar().ifPresent(ServerBossBar::deregisterPlayers);
+                            v.getEntities().forEach(e -> e.setGravity(true));
                         }
-                    } catch (Throwable e2) {
-                        context.getBar().ifPresent(ServerBossBar::deregisterPlayers);
-                        v.getEntities().forEach(e -> e.setGravity(true));
-                    }
+                    });
                 }
             });
-        }catch (ConcurrentModificationException ignore){
+        } catch (ConcurrentModificationException ignore) {
 
         }
     }
 
-    public static Scheduler createScheduler(){
+    public static Scheduler createScheduler() {
         return CorePlugin.createSchedulerBuilder()
                 .setExecutor(new FallExecutor())
                 .setIterationUnit(ShipsPlugin.getPlugin().getConfig().getFallingDelayUnit())
